[CMake] PR 26385 Add some checks for DYLIB settings
PR 26385 reported issues linking tools against the LLVM dynamic library. The root problem is that CMake's caching behavior goes really wonky with settings that have default values influenced by other settings. This patch just adds some tests at the end of the block to make sure the options are used in a valid combination.

This preserves changing the default value based on the specified values which is convenient, but I think we should probably simplify this by just defaulting LLVM_DISABLE_LLVM_DYLIB_ATEXIT to Off as it has very few clients, and none of them in-tree.